Support Windows Subsystem for Linux (WSL)
WSL is a proper Linux distro so `process.platform === 'linux'`. However, most people won't have an X server installed so `xsel` won't work.
But WSL can call native windows executables and the fallback/clipboard_*.exe programs work fine.
Thus, if we use the windows lib on WSL, everything just works.
Code for checking whether we're running in WSL is based on https://stackoverflow.com/a/43618657, which has a quote from a WSL dev as the source.